fix(conversion): correct conversion precision

+++ src/main.rs
@@ -1,7 +1,7 @@
 fn main() {
-    let celcius: i32 = 36;
+    let celcius: f64 = 36.0;
 
-    let fahrenheit = ( 9 / 5 * celcius) + 32;
+    let fahrenheit = (9.0 / 5.0 * celcius) + 32.0;
 
     println!("Suhu dalam celcius: {celcius}");
     println!("Suhu dalam fahrenheit: {fahrenheit}");
